Remove redundant functor
